:recycle: Refactoring code.
