+++ main/src/components/AppInstallLoadingPanel/progress.js
