+++ main/src/components/AppSetting/index.js
@@ -1,0 +1,10 @@
+/*
+ * @LastEditors: zhanghengxin ezreal.zhang@icewhale.org
+ * @LastEditTime: 2023/8/22 下午5:41
+ * @FilePath: /CasaOS-UI/main/src/components/AppSetting/index.js
+ * @Description:
+ *
+ * Copyright (c) 2023 by IceWhale, All Rights Reserved.
+
+ */
+
