+++ main/src/components/Apps/AddNewButtonCard.vue
@@ -4,9 +4,9 @@
  * @LastEditors: Jerryk jerry@icewhale.org
  * @LastEditTime: 2022-06-22 18:26:23
  * @FilePath: /CasaOS-UI/src/components/Apps/AddNewButtonCard.vue
- * @Description: 
- * 
- * Copyright (c) 2022 by 老竭力 jerrykuku@qq.com, All Rights Reserved. 
+ * @Description:
+ * lost.
+ * Copyright (c) 2022 by 老竭力 jerrykuku@qq.com, All Rights Reserved.
 -->
 <template>
 	<div class="column is-narrow is-3">
@@ -15,7 +15,7 @@
 
 			<div class="cards-content">
 				<!-- Card Content Start -->
-				<b-tooltip :label="$t('Open')" type="is-dark" animation="fade1" :animated="true">
+				<b-tooltip :animated="true" :label="$t('Open')" animation="fade1" type="is-dark">
 					<div
 					class="has-text-centered is-flex is-justify-content-center is-flex-direction-column pt-3 pb-3 img-c">
 						<a class="is-flex is-justify-content-center">
