+++ main/src/components/Apps/AppSection.vue
@@ -52,7 +52,6 @@
 				</div>
 			</template>
 			<!-- App Icon Card End -->
-			<!-- <b-loading slot="footer" v-model="isLoading" :is-full-page="false"></b-loading> -->
 		</draggable>
 		<!-- App List End -->
